requirements.txt
Django>=4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
pymongo>=3.12.0,<4.0.0
dnspython>=2.1.0
django-cors-headers>=3.14.0
